Fix link between color input and other color changes
